Convert static libcommon library to dynamic libceph_common library
libcommon as a static library causes issues with LTTng, because it ends
up linked statically into both executables and dynamic libraries.
When the executables load the dynamic libraries, the same tracepoints
get loaded twice, and LTTng aborts the process.  This is why libcommon
was changed to a dynamic library.

libcommon was renamed to libceph_common because it will now be an
installed library, and the name libcommon is too generic and prone to
clashes.  The new installed library will probably require packaging
changes in the future.

Many test outputs (*.t) were modified to change "toolname: <msg>" to
"*toolname: <msg> (glob)".  Now that libcommon is a dynamic library,
many tools which previously did not rely on any Ceph dynamic libraries
changed from being invoked directly in tests to being invoked through a
libtool-generated script, which changed argv[0] in the program, which
changed the program's output.

The global variable coll_t::META_COLL was removed because its destructor
was called twice when it was linked into an executable as well as
libcommon, causing a segfault.

Signed-off-by: Adam Crume <adamcrume@gmail.com>